+++ NLP/Scoring/SimpleUnigram/Default.hs
@@ -12,5 +12,5 @@ import NLP.Scoring.SimpleUnigram.Import
 -- | Default simple unigram scores for a system of consonants, liquid
 -- consonants, and vowels of arbitrary scale.
 
-clvDefaults = genSimpleScoring $ decodeUtf8 $(embedFile "scoring/simpleunigram.txt")
+clvDefaults = genSimpleScoring $ decodeUtf8 $(embedFile "scoring/simpleunigram.score")